requirements.txt
streamlit
pdfplumber
python-docx
pandas
extract-msg
pillow
pytesseract
google-generativeai